Move aux files to model dir
