+++ c/ldv-sets/test_mutex_double_lock_false-unreach-call.c
@@ -1,7 +1,7 @@
 #include <stdlib.h>
 
-#include "ldv.h"
-#include "mutex.c"
+#include "model/ldv.h"
+#include "model/mutex.c"
 
 void foo()
 {
